_project-fullstack/mcp-server/requirements.txt
mcp>=1.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
qdrant-client>=1.9.0
supabase>=2.4.0
openai>=1.30.0
python-dotenv>=1.0.0
